requirements.txt
Django==1.6.5
Markdown==2.4.1
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
cmsplugin-filer==0.9.9
django-ace==1.0.1
django-admin-bootstrapped==1.6.5
django-appconf==0.6
django-classy-tags==0.5.1
django-cms==3.0.2
django-email-obfuscator==0.1.2
django-filebrowser==3.5.6
django-filer==0.9.5
django-grappelli==2.5.3
django-markupfield==1.2.0
django-mptt==0.6.1
django-polymorphic==0.5.5
django-sekizai==0.7
djangocms-admin-style==0.2.2
djangocms-cascade==0.3.1
djangocms-file==0.0.1
djangocms-flash==0.0.2
djangocms-text-ckeditor==2.1.6
docutils==0.11
easy-thumbnails==2.0.1
html5lib==0.999
jsonfield==0.9.20
psycopg2==2.5.3
six==1.7.2
sqlparse==0.1.11
textile==2.1.8
trans==2.0.1
wsgiref==0.1.2
